+++ crypto/x509/name_print.cc
@@ -136,7 +136,7 @@ static int do_name_ex(BIO *out, const X509_NAME *n, int indent,
     } else {
       objbuf = OBJ_nid2sn(fn_nid);
     }
-    int objlen = strlen(objbuf);
+    int objlen = (int)strlen(objbuf);
     if (!maybe_write(out, objbuf, objlen) ||
         !maybe_write(out, sep_eq, sep_eq_len)) {
       return -1;
