+++ ssl/ssl_session.cc
@@ -369,7 +369,7 @@ static int ssl_encrypt_ticket_with_cipher_ctx(SSL_HANDSHAKE *hs, CBB *out,
   } else {
     int len;
     if (!EVP_EncryptUpdate(ctx.get(), ptr + total, &len, session_buf,
-                           session_len)) {
+                           (int)session_len)) {
       return 0;
     }
     total += len;
@@ -542,7 +542,7 @@ static enum ssl_hs_wait_t ssl_lookup_session(
   if (!session && ssl->session_ctx->get_session_cb != nullptr) {
     int copy = 1;
     session.reset(ssl->session_ctx->get_session_cb(ssl, session_id.data(),
-                                                   session_id.size(), &copy));
+                                                   (int)session_id.size(), &copy));
     if (!session) {
       return ssl_hs_ok;
     }
@@ -844,7 +844,7 @@ void SSL_SESSION_free(SSL_SESSION *session) {
 const uint8_t *SSL_SESSION_get_id(const SSL_SESSION *session,
                                   unsigned *out_len) {
   if (out_len != NULL) {
-    *out_len = session->session_id.size();
+    *out_len = (int)session->session_id.size();
   }
   return session->session_id.data();
 }
@@ -937,7 +937,7 @@ uint32_t SSL_SESSION_set_timeout(SSL_SESSION *session, uint32_t timeout) {
 const uint8_t *SSL_SESSION_get0_id_context(const SSL_SESSION *session,
                                            unsigned *out_len) {
   if (out_len != NULL) {
-    *out_len = session->sid_ctx.size();
+    *out_len = (int)session->sid_ctx.size();
   }
   return session->sid_ctx.data();
 }
